requirements.txt
asgiref==3.4.1
certifi==2021.10.8
Django==4.0.1
django-cors-headers==3.11.0
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.2
django-filter==21.1
django-nine==0.2.5
djangorestframework==3.13.1
elasticsearch==7.16.3
elasticsearch-dsl==7.4.0
Markdown==3.3.6
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
